LayoutTests/inspector/debugger/breakpoint-manager-listeners-count-expected.txt

Tests that scripts panel does not create too many source frames.

Debugger was enabled.

Running: testSourceFramesCount
Page reloaded.
Number of breakpoint-added event listeners is 3
Dumping SourceFrames listening for breakpoint-added event:
    script1.js
    script2.js
Debugger was disabled.
